Update smoke tests
